Checking for isDebugEnabled was added to CourseServiceImpl and GroupServiceImpl. Signed-off-by: VasyaDen <vasyaden@mail.ru>

+++ src/main/java/com/foxminded/foxuniversity/service/implementation/CourseServiceImpl.java
@@ -29,7 +29,9 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public List<Course> getAll() {
-        logger.info("CourseService calls courseDao.getAll().");
+        if (logger.isDebugEnabled()) {
+            logger.info("CourseService calls courseDao.getAll().");
+        }
         List<Course> courses = courseDao.getAll();
         fillLessons(courses);
         fillGroups(courses);
@@ -38,7 +40,9 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public Course getById(int id) {
-        logger.info("CourseService calls courseDao.getById(" + id + ").");
+        if (logger.isDebugEnabled()) {
+            logger.info("CourseService calls courseDao.getById(" + id + ").");
+        }
         Course course = courseDao.getById(id);
         fillLessons(course);
         fillGroups(course);
@@ -47,7 +51,9 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public List<Course> getByGroup(Group group) {
-        logger.debug("CourseService calls courseDao.getByGroup(" + group + ").");
+        if (logger.isDebugEnabled()) {
+            logger.debug("CourseService calls courseDao.getByGroup(" + group + ").");
+        }
         List<Course> courses = courseDao.getByGroup(group);
         fillLessons(courses);
         fillGroups(courses);
@@ -56,22 +62,30 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public void save(Course course) {
-        logger.debug("CourseService calls courseDao.save(Course{id = " + course.getId() + "}).");
+        if (logger.isDebugEnabled()) {
+            logger.debug("CourseService calls courseDao.save(Course{id = " + course.getId() + "}).");
+        }
         courseDao.save(course);
     }
 
     @Override
     public boolean update(Course course) {
-        logger.debug("CourseService calls courseDao.update(Course{id = " + course.getId() + "}).");
+        if (logger.isDebugEnabled()) {
+            logger.debug("CourseService calls courseDao.update(Course{id = " + course.getId() + "}).");
+        }
         return courseDao.update(course);
     }
 
     @Override
     public boolean delete(Course course) {
-        logger.debug("Checking for teachers in the Course{id = " + course.getId() + "}).");
+        if (logger.isDebugEnabled()) {
+            logger.debug("Checking for teachers in the Course{id = " + course.getId() + "}).");
+        }
         if (teacherService.getByCourse(course).isEmpty()) {
-            logger.debug("There are no teachers in the course. " +
-                    "Call courseDao.delete(Course{id = " + course.getId() + "}).");
+            if (logger.isDebugEnabled()) {
+                logger.debug("There are no teachers in the course. " +
+                        "Call courseDao.delete(Course{id = " + course.getId() + "}).");
+            }
             return courseDao.delete(course);
         }
         logger.warn("There are some teachers in the Course{id = " + course.getId() + "}). Deletion canceled.");
@@ -79,7 +93,9 @@ public class CourseServiceImpl implements CourseService {
     }
 
     private void fillLessons(Course course) {
-        logger.debug("Set lessons to course: call lessonService.getByCourse(Course{id = " + course.getId() + "}).");
+        if (logger.isDebugEnabled()) {
+            logger.debug("Set lessons to course: call lessonService.getByCourse(Course{id = " + course.getId() + "}).");
+        }
         course.setLessons(lessonService.getByCourse(course));
     }
 
@@ -90,7 +106,9 @@ public class CourseServiceImpl implements CourseService {
     }
 
     private void fillGroups(Course course) {
-        logger.debug("Set lessons to course: call groupService.getByCourse(Course{id = " + course.getId() + "}).");
+        if (logger.isDebugEnabled()) {
+            logger.debug("Set lessons to course: call groupService.getByCourse(Course{id = " + course.getId() + "}).");
+        }
         course.setGroups(groupService.getByCourse(course));
     }
